gemini_ocr/requirements.txt
google-generativeai
pdf2image
pillow
python-dotenv